(1) Functions will be be delegated even when present in the scope an explicit BEGIN/COMMIT transaction block or UDF calling another UDF. (2) Prohibit/Limit the delegated function not to do a 2PC (or any work on a remote connection) (3) Have a safety net around ensuring the (2) i.e. we should block the connections from the delegated procedure or make sure that no 2PC happens on the node. (4) Such delegated functions should only use the distributed argument value
To limit the scope of the project we are considering only Functions(not procedures) for the initial work.

DESCRIPTION: Introduce a new flag for "forcePushdown" in create_distributed_function(), which will allow the function to be delegated in an explicit transaction block.

Fixes #3265

Once the function is delegated to the worker, on that node during the planning

distributed_planner()
   TryToDelegateFunctionCall()
      CheckDelegatedFunctionExecution()
         EnableInForceDelegatedFuncExecution() Save the distribution argument (Constant)
ExecutorStart()
  CitusBeginScan()
    IsShardKeyValueAllowed()  Make sure to not use non-distribution argument"

ExecutorRun()
  AdaptiveExecutor()
    StartDistributedExecution()
      EnsureDelegatedFunctionSafeExecution()  Make sure all the shards are local to the node in remoteTaskList
This PR includes changes from #5060

Triggers
Standalone SQL delegated to a worker node and trigger-function fires on that shard.
Function(force-pushdown) delegated to a worker node and trigger fires for the individual SQL(s) within the body of the function.
For the scenario(1), the SQL shard key must be saved during the SQL execution, and when trigger function fires, it's workerJob->partitionKeyValue must be checked to match the saved value.

For the scenario(2), we must ensure that Triggers' should act only on the distributed argument of the function.